refactor(cli): fold provider-hooks into apps/cli ingress
The package's only remaining job after observer-side normalization was
the writer: stdin parse, raw event delivery, spool writes, auto-start.
That now lives in apps/cli/src/ingress with the stn-ingress bin built
as apps/cli/dist/ingressMain.js (repo bin/stn-ingress retargeted;
installed hook scripts call stn-ingress by name and need no change).
Spool schemas were already in contracts; the observer drainer already
lived in apps/observer. Deletes packages/provider-hooks.
