+++ apps/cli/package.json
@@ -11,10 +11,15 @@
     "./internal": {
       "types": "./dist/internal.d.ts",
       "import": "./dist/internal.js"
+    },
+    "./ingress": {
+      "types": "./dist/ingress/index.d.ts",
+      "import": "./dist/ingress/index.js"
     }
   },
   "bin": {
-    "station": "./dist/main.js"
+    "station": "./dist/main.js",
+    "stn-ingress": "./dist/ingressMain.js"
   },
   "scripts": {
     "build": "tsc -p tsconfig.json",
