+++ apps/cli/src/ingress/command.ts
@@ -359,7 +359,7 @@ function parsePositiveInteger(value: string, flag: string): number {
 }
 
 function defaultObserverEntryPath(): string {
-  return fileURLToPath(new URL("../../../apps/cli/dist/observerMain.js", import.meta.url));
+  return fileURLToPath(new URL("../observerMain.js", import.meta.url));
 }
 
 function formatRejectedReceipt(receipt: ProviderHookReceipt): string {
