+++ apps/cli/src/ingress/stdin.ts
