--- packages/provider-hooks/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@station/provider-hooks",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "bin": {
-    "stn-ingress": "./dist/main.js"
-  },
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "typecheck": "tsc -p tsconfig.json --noEmit"
-  },
-  "dependencies": {
-    "@station/claude": "workspace:*",
-    "@station/codex": "workspace:*",
-    "@station/config": "workspace:*",
-    "@station/contracts": "workspace:*",
-    "@station/cursor": "workspace:*",
-    "@station/observability": "workspace:*",
-    "@station/pi": "workspace:*",
-    "@station/protocol": "workspace:*",
-    "@station/runtime": "workspace:*",
-    "@station/worktrunk": "workspace:*"
-  }
-}
