--- packages/provider-hooks/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "tsBuildInfoFile": "dist/tsconfig.tsbuildinfo",
-    "paths": {
-      "@station/claude": ["../../integrations/harness/claude/dist/index.d.ts"],
-      "@station/codex": ["../../integrations/harness/codex/dist/index.d.ts"],
-      "@station/config": ["../config/dist/index.d.ts"],
-      "@station/contracts": ["../contracts/dist/index.d.ts"],
-      "@station/cursor": ["../../integrations/harness/cursor/dist/index.d.ts"],
-      "@station/observability": ["../observability/dist/index.d.ts"],
-      "@station/pi": ["../../integrations/harness/pi/dist/index.d.ts"],
-      "@station/protocol": ["../protocol/dist/index.d.ts"],
-      "@station/runtime": ["../runtime/dist/index.d.ts"],
-      "@station/worktrunk": ["../../integrations/worktree/worktrunk/dist/index.d.ts"]
-    }
-  },
-  "include": ["src/**/*.ts"]
-}
